requirements.txt
flask==2.3.3
RPi.GPIO==0.7.1